Rollback auto-unapprove since it relies on a plugin
The recently added (see #25) ability to auto-unapprove changes seems
to rely on an Atlassian plugin:

https://docs.atlassian.com/bitbucket-server/rest/7.10.0/bitbucket-rest.html#idp409
https://marketplace.atlassian.com/apps/1211449/auto-unapprove-for-bitbucket-server?hosting=server&tab=overview

For this reason, we cannot include it in the current form, since it
would cause BEC to fail on systems which do not have the plugin
installed.

We can try adding it back, but as an optional argument.