Fix Dialyzer error
REST API docs for the

PUT /rest/api/1.0/projects/{projectKey}/repos/{repositorySlug}/branches/default

endpoint says that the body should be a map '{"id":
branchid}'. Dialyzer was warning about the third argument to
bitbucket_api:set_default_branch/3 being a binary instead of a
map. Apparently this works anyway, but we pass a map instead, so that
we both satisfy Dialyzer and follow the REST API docs.